Interim commit, working towards storing incidents to DB.
- Updated struckTOBO with new public methods to generate tweet and
  incident VOs
- Added test for new incident generate method
- Updated IncidentVO to reflect DB
- Added stub for StruckTOTweetDAO

+++ src/ca/gkwb/struckto/tweet/StruckTOTweetVO.java
@@ -11,7 +11,7 @@ public class StruckTOTweetVO {
 	
 //	"TWEET_ID" integer NOT NULL PRIMARY KEY,
 	@Getter @Setter
-	private int tweetId;
+	private long tweetId;
 //	"TWEET_URL" character varying,
 	@Getter @Setter
 	private String url;
